requirements.txt
cffi==1.17.1
cryptography==43.0.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.5.0
more-itertools==10.5.0
pycparser==2.22
pycryptodome==3.21.0
SecretStorage==3.3.3
